Initial work on fixing Clang Static analyzer warnings in regards to allocation.

+++ fontforge/nowakowskittfinstr.c
@@ -295,7 +295,7 @@ int TTF__getcvtval(SplineFont *sf,int val) {
         cvt_tab = chunkalloc(sizeof(struct ttf_table));
         cvt_tab->tag = CHR('c','v','t',' ');
         cvt_tab->maxlen = 200;
-        cvt_tab->data = malloc(100*sizeof(short));
+        cvt_tab->data = malloc(100*sizeof(uint8));
         cvt_tab->next = sf->ttf_tables;
         sf->ttf_tables = cvt_tab;
     }
@@ -727,7 +727,7 @@ static void init_cvt(GlobalInstrCt *gic) {
     }
 
     cvtsize = cvtindex;
-    cvt = realloc(cvt, cvtsize * sizeof(int16));
+    cvt = realloc(cvt, cvtsize * sizeof(uint8));
 
     /* Try to implant the new cvt table */
     gic->cvt_done = 0;
